de-linting after averaging & ProviderTimer class devel
Grainfather Custom URL provider now operates on a machine.Timer. This Timer has an adjust(able ) method that allows for different intervals to be set., falling back to the default_interval after the timer interrupt.
Global and per provider averaging period can be set, or use latest value (if it is within log period)