add a forced log file write
logging has been moved into a module.
The RotatingLogFileHandler has been overridden with TimedRotatingLogFileHandler
TimedRotatingLogFileHandler requires an additional parameter which is the forced write interval in seconds. Currently hardcoded to 5mins for debug and 15mins for csv logs. If the device is shut down by pulling power then up to x mins of each log will not be available. Previously this was 4096 bytes. If each record is 36bytes then that would represent 113 log points lost. This should allow for a balance between hammering the flash memory and keeping information up to date