remove wifi_check interval
wifi check interval in config was set to 5 mins by default. This is now hardcoded at 50 secs. When wifi is connected the check is simple and quick, adding a configurable variable adds nothing. 50 secs is a reasonable check rate when logging is expected to be of the order of every 3600 secs.
bump version